adjust construction
